read-file-app/src/infrastructure/kafkaproducer.py
import json
from kafka import KafkaProducer

class ProducerKafka:
    broker = ""
    topic = ""
    producer = None

    def __init__(self, broker, topic):
        self.broker = broker
        self.topic = topic
        self.producer = KafkaProducer(bootstrap_servers=self.broker,
        value_serializer=lambda v: json.dumps(v).encode('utf-8'),
        acks='all',
        retries = 3)


    def send_msg(self, msg):
        print("sending message...", flush=True)
        try:
            future = self.producer.send(self.topic, msg)
            self.producer.flush()
            future.get(timeout=60)
            print("message sent successfully...", flush=True)
            return {'status_code':200, 'error':None}
        except Exception as ex:
            print(ex, flush=True)