requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.2
celery==5.3.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.2.0
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
kombu==5.3.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.25.2
pandas==2.1.0
prompt-toolkit==3.0.39
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
